@@ -1,24 +1,24 @@
 name: sign_in_with_apple
 description: Flutter bridge to initiate Sign in with Apple (on iOS, macOS, and Android). Includes support for keychain entries as well as signing in with an Apple ID.
-version: 2.5.1
+version: 3.1.0
 homepage: https://github.com/aboutyou/dart_packages/tree/master/packages/sign_in_with_apple
 repository: https://github.com/aboutyou/dart_packages
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.8 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
-  meta: ^1.0.0
-
   flutter:
     sdk: flutter
 
-dev_dependencies:
-  pedantic: "1.8.0+1"
+  meta: ^1.3.0
 
+dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  pedantic: ^1.11.0
 
 flutter:
   plugin:
