@@ -8,8 +8,8 @@ homepage: https://github.com/aboutyou/dart_packages/tree/master/packages/with_bl
 repository: https://github.com/aboutyou/dart_packages
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
-  flutter: ">=1.7.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
@@ -18,12 +18,13 @@ dependencies:
   meta: ^1.0.0
 
 dev_dependencies:
-  test: any
+  flutter_test:
+    sdk: flutter
 
   state_queue:
     path: ../../state_queue
+
+  test: any
+
   with_bloc:
     path: ./..
-
-  flutter_test:
-    sdk: flutter
